requirements.txt
streamlit==1.25.0
Pillow==9.5.0
openai==1.3.0
python-dotenv==0.10.1